auth: refactor
